backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-dotenv==1.0.0
ollama==0.1.6
chromadb==0.4.18

# Telegram bot integration
python-telegram-bot==21.0
APScheduler==3.10.4

# Enhanced Obsidian features
python-frontmatter==1.1.0
watchdog==4.0.0
aiofiles==23.2.1

# Claude API (alternative to Ollama)
anthropic==0.18.1
